Refactored MockStorageFactory to InMemoryDatabase
