+++ packages/sequencer/src/index.ts
@@ -37,8 +37,10 @@ export * from "./protocol/production/unproven/UnprovenProducerModule";
 export * from "./storage/model/Block";
 export * from "./storage/repositories/BlockStorage";
 export * from "./storage/repositories/UnprovenBlockStorage";
-export * from "./storage/MockStorageDependencyFactory";
-export * from "./storage/Database";
+export * from "./storage/inmemory/InMemoryDatabase";
+export * from "./storage/inmemory/InMemoryAsyncMerkleTreeStore";
+export * from "./storage/inmemory/InMemoryBlockStorage";
+export * from "./storage/inmemory/InMemoryBatchStorage";
 export * from "./storage/StorageDependencyFactory";
 export * from "./helpers/query/QueryTransportModule";
 export * from "./helpers/query/QueryBuilderFactory";
