--- packages/sequencer/src/storage/Database.ts
@@ -1,2 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-empty-interface
-export interface Database {}
